Changing macro names
Changing the following macro names:

unknownsfx0x20 to SquareSound - This command, when used in an .asm file that defines an audio piece, generates a square tone modulated by 4 values.
unknownnoise0x20 to NoiseSound - This command is similar to SquareSound, but it uses 3 values instead of 4 and generates noise instead of a square tone.
unknownsfx0x10 to eSquarePitchEnvelope - This command is not an unknown sound effect, but a modifier for the above two commands which it precedes; it takes one decimal value which is translated into a hexadecimal value when used by the system, and depending on the second digit of this hex value the sounds generated by SquareSound and NoiseSound will have their frequency slide either up (1-7) or down (9-F) or not at all (0,8). E.g. typing 23 will have the system read $17, which will modulate the frequency to slide up because the second digit is 7.